fix loop import
